Added a screen to see a match.
