+++ aston_project_3/apps/game/views/__init__.py
@@ -1,6 +1,7 @@
 """The module file."""
 from .bot_deleted_view import BotDeletedViewSet
 from .bot_matches_view import BotMatchesViewSet
+from .bot_match_data_view import BotMatchDataViewSet
 from .bot_match_view import BotMatchViewSet
 from .create_bot_view import CreateBotViewSet
 from .delete_bot_view import DeleteBotViewSet
