+++ aston_project_3/static/style/game/match/match.css
@@ -2,6 +2,7 @@
     width: 100%;
     /* height: 100%; */
     display: block;
+    border: 1px solid var(--grey-color);
 }
 
 .control-panel {
