+++ aston_project_3/templates/base.html
@@ -98,7 +98,7 @@
                                     <a
                                         class="dropdown-item link-light dropdown-nav-item"
                                         href="/bots">
-                                        {% translate "see_matches" %}
+                                        {% translate "see_my_matches" %}
                                     </a>
                                 </li>
                                 <li>
