+++ aston_project_3/templates/game/create_bot.html
@@ -68,15 +68,6 @@
                             <em class="fas fa-forward-fast fa-lg"></em>
                         </button>
                     </div>
-                    {% comment %} <div class="progress">
-                        <div
-                            class="progress-bar"
-                            role="progressbar"
-                            aria-valuenow="0"
-                            aria-valuemin="0"
-                            aria-valuemax="100">
-                        </div>
-                    </div> {% endcomment %}
                 </div>
                 <div id="fight-result" class="d-none text-color-fore"></div>
                 <div id="block-output" class="p-3 mt-1 back-color-white-veil-5 d-none"></div>
